feat(AppShell): 实现响应式布局并集成AppShell组件
- 添加shadcn组件配置文件(components.json)
- 集成AppShell组件以支持响应式侧边栏和抽屉功能
- 实现移动端适配的会话面板抽屉切换逻辑
- 添加媒体查询钩子用于响应式断点检测
- 更新键盘快捷键行为以适应不同屏幕尺寸
- 重构App组件结构以使用新的布局系统
