+++ Frontend/src/components/ui-shadcn/sheet.tsx
@@ -1,0 +1,95 @@
+import * as DialogPrimitive from "@radix-ui/react-dialog";
+import { X } from "lucide-react";
+import { forwardRef } from "react";
+import { cn } from "@/lib/util";
+
+export const Sheet = DialogPrimitive.Root;
+export const SheetTrigger = DialogPrimitive.Trigger;
+export const SheetClose = DialogPrimitive.Close;
+export const SheetPortal = DialogPrimitive.Portal;
+export const SheetTitle = DialogPrimitive.Title;
+export const SheetDescription = DialogPrimitive.Description;
+
+export const SheetOverlay = forwardRef<
+  HTMLDivElement,
+  React.ComponentPropsWithoutRef<typeof DialogPrimitive.Overlay>
+>(({ className, ...props }, ref) => (
+  <DialogPrimitive.Overlay
+    ref={ref}
+    className={cn(
+      "fixed inset-0 z-50 bg-ink-950/35 backdrop-blur-[1px]",
+      "data-[state=open]:animate-dialog-in data-[state=closed]:animate-dialog-out",
+      className,
+    )}
+    {...props}
+  />
+));
+SheetOverlay.displayName = "SheetOverlay";
+
+type SheetSide = "left" | "right";
+
+const sideClasses: Record<SheetSide, string> = {
+  left: "left-0 border-r data-[state=open]:animate-dialog-in",
+  right: "right-0 border-l data-[state=open]:animate-dialog-in",
+};
+
+export interface SheetContentProps
+  extends React.ComponentPropsWithoutRef<typeof DialogPrimitive.Content> {
+  side?: SheetSide;
+  title?: string;
+  description?: string;
+  showClose?: boolean;
+}
+
+export const SheetContent = forwardRef<HTMLDivElement, SheetContentProps>(
+  ({ side = "right", className, children, title, description, showClose = true, ...props }, ref) => (
+    <SheetPortal>
+      <SheetOverlay />
+      <DialogPrimitive.Content
+        ref={ref}
+        className={cn(
+          "fixed top-0 z-50 flex h-full w-[min(360px,calc(100vw-24px))] flex-col overflow-hidden",
+          "border-ink-200 bg-paper-50 shadow-soft outline-none",
+          "data-[state=closed]:animate-dialog-out",
+          sideClasses[side],
+          className,
+        )}
+        {...props}
+      >
+        {(title || showClose) ? (
+          <div className="flex min-h-14 items-start gap-3 border-b border-ink-200/70 bg-paper-50 px-4 py-3.5">
+            <div className="min-w-0 flex-1">
+              {title ? (
+                <DialogPrimitive.Title className="truncate text-[13.5px] font-medium text-ink-900">
+                  {title}
+                </DialogPrimitive.Title>
+              ) : null}
+              {description ? (
+                <DialogPrimitive.Description className="mt-0.5 truncate text-[12px] text-ink-500">
+                  {description}
+                </DialogPrimitive.Description>
+              ) : null}
+            </div>
+            {showClose ? (
+              <DialogPrimitive.Close asChild>
+                <button
+                  type="button"
+                  className={cn(
+                    "grid h-8 w-8 shrink-0 place-items-center rounded-md text-ink-500 transition",
+                    "hover:bg-ink-900/[0.05] hover:text-ink-800",
+                    "focus:outline-none focus:ring-2 focus:ring-terra-200/60",
+                  )}
+                  aria-label="关闭"
+                >
+                  <X className="h-4 w-4" />
+                </button>
+              </DialogPrimitive.Close>
+            ) : null}
+          </div>
+        ) : null}
+        {children}
+      </DialogPrimitive.Content>
+    </SheetPortal>
+  ),
+);
+SheetContent.displayName = "SheetContent";
